search is working, asynchronous with json return and js print, cqn extend following the model shown in the examples

+++ src/housing/models.py
@@ -6,7 +6,7 @@ from django.contrib.auth.models import User
 class House(models.Model):
     name = models.CharField(max_length=30, verbose_name="Name")
     surface = models.PositiveSmallIntegerField(verbose_name="Surface")
-    price = models.DecimalField(max_digits=8, decimal_places=2, verbose_name="Price")
+    price = models.PositiveSmallIntegerField(verbose_name="Price")
     
     def __unicode__(self):
         return u"%s"%self.name
